+++ Railgun/Commands/Inactivity/InactivityKick.cs
@@ -1,0 +1,11 @@
+﻿using Finite.Commands;
+using Railgun.Core;
+
+namespace Railgun.Commands.Inactivity
+{
+    public partial class Inactivity
+    {
+        [Alias("kick")]
+        public partial class Kick : SystemBase { }
+    }
+}
